// src/app/enums/weather-type.enum.ts
/**
 * 天气状况
 */
export enum WeatherType {
  /**	晴	1 */
  Sunny = 1,
  /**	多云	2 */
  Cloudy = 2,
  /**	阴	3 */
  Overcast = 3,
  /**	阵雨	4 */
  Shower = 4,
  /**	雷阵雨	5 */
  Thunder = 5,
  /**	雨夹雪	7 */
  Sleet = 6,
  /**	小雨	8 */
  LightRain = 7,
  /**	冰雹	6 */
  Hailstone = 6,
  /**	中雨	9 */
  ModerateRain = 9,
  /**	大雨	10 */
  HeavyRain = 10,
  /**	暴雨	11 */
  Rainstorm = 11,
  /**	阵雪	12 */
  Snow = 12,
}